+++ openghg/util/_function_inputs.py
@@ -1,7 +1,8 @@
 from typing import Any
 from collections.abc import Callable
 import logging
-import inspect
+
+from openghg.util._registry import get_parameters
 
 logger = logging.getLogger("openghg.util.function_inputs")
 logger.setLevel(logging.DEBUG)  # Have to set level for logger as well as handler
@@ -24,8 +25,7 @@ def split_function_inputs(
         dict, dict: Dictionaries for parameters accepted and not accepted by the function
     """
     # Find parameters that fn accepts
-    signature = inspect.signature(fn)
-    fn_accepted_parameters = [param.name for param in signature.parameters.values()]
+    fn_accepted_parameters = get_parameters(fn)
 
     fn_parameters = {}
     remaining_parameters = {}